Updated card footer background color
The left border was weirdly mixing with the page background. This new color
helps it stand out from the background, makes the card footer text a bit more readable,
and doesn't distract from the rest of the card or page.

Also added note about where the color scheme comes from.